fix dtype mismatch

+++ src/transformers/models/qianfan_ocr/modeling_qianfan_ocr.py
@@ -167,7 +167,7 @@ class QianfanViTEmbeddings(nn.Module):
 
     def forward(self, pixel_values: torch.FloatTensor, bool_masked_pos=None, **kwargs) -> torch.Tensor:
         target_dtype = self.patch_embedding.weight.dtype
-        patch_embeds = self.patch_embedding(pixel_values)
+        patch_embeds = self.patch_embedding(pixel_values.to(target_dtype))
         batch_size, _, height, width = patch_embeds.shape
         patch_embeds = patch_embeds.flatten(2).transpose(1, 2)
         class_embeds = self.class_embedding.expand(batch_size, 1, -1).to(target_dtype)
